Add more DNC actions, split count sim into separate file
