+++ packages/frontend/src/scripts/sims/ranged/dnc_sim.ts
@@ -1,30 +1,10 @@
-import {
-    Ability,
-    CombinedBuffEffect,
-    GcdAbility,
-    PartyBuff,
-    SimResult,
-    SimSpec,
-    Simulation
-} from "@xivgear/core/sims/sim_types";
-import {BuffSettingsExport, BuffSettingsManager} from "@xivgear/core/sims/common/party_comp_settings";
-import {defaultResultSettings, ResultSettings} from "@xivgear/core/sims/cycle_sim";
-import {addValues, applyStdDev, multiplyFixed, ValueWithDev} from "@xivgear/xivmath/deviation";
+import {AutoAttack, GcdAbility, OgcdAbility, SimSpec} from "@xivgear/core/sims/sim_types";
 import {CharacterGearSet} from "@xivgear/core/gear";
-import {JobName} from "@xivgear/xivmath/xivconstants";
-import {NamedSection} from "../../components/section";
-import {BuffSettingsArea} from "../party_comp_settings";
-import {ResultSettingsArea} from "../components/result_settings";
-import {writeProxy} from "@xivgear/core/util/proxies";
-import {simpleAutoResultTable} from "../components/simple_tables";
-import {quickElement} from "../../components/util";
-import {abilityEquals} from "@xivgear/core/sims/ability_helpers";
-import {abilityToDamageNew, combineBuffEffects, noBuffEffects} from "@xivgear/core/sims/sim_utils";
-import {CustomTable, HeaderRow} from "../../tables";
+import {BaseUsageCountSim, CountSimResult, ExternalCountSettings, SkillCount} from "../processors/count_sim";
 
 export const dncDtSheetSpec: SimSpec<DncDtSim, DncDtSimSettings> = {
     displayName: "DNC Level 100 Sim",
-    loadSavedSimInstance(exported: DncDtSimSettingsExport) {
+    loadSavedSimInstance(exported: ExternalCountSettings<DncDtSimSettings>) {
         return new DncDtSim(exported);
     },
     makeNewSimInstance(): DncDtSim {
@@ -38,25 +18,15 @@ export const dncDtSheetSpec: SimSpec<DncDtSim, DncDtSimSettings> = {
 
 export type DncDtSimSettings = NonNullable<unknown>
 
-export type DncDtSimSettingsExport = {
-    customSettings: DncDtSimSettings,
-    buffConfig: BuffSettingsExport,
-    resultSettings: ResultSettings,
-}
 
-export interface DncDtSimResults extends SimResult {
-    unbuffedPps: number,
-    totalDamage: ValueWithDev,
-    mainDpsFull: ValueWithDev,
-    cycleTime: number,
-    buffBuckets: BuffWindowUsages[]
+export interface DncDtSimResults extends CountSimResult {
 }
 
 const cascade: GcdAbility = {
     name: 'Cascade',
     type: 'gcd',
     attackType: 'Weaponskill',
-    potency: 220,
+    potency: 280,
     gcd: 2.5,
     id: 15989
 } as const satisfies GcdAbility;
@@ -65,136 +35,187 @@ const fountain: GcdAbility = {
     name: 'Fountain',
     type: 'gcd',
     attackType: 'Weaponskill',
-    potency: 280,
+    potency: 340,
     gcd: 2.5,
     id: 15990
 } as const satisfies GcdAbility;
 
-type SkillCount = [ability: Ability, count: number];
 
-type BuffWindowUsages = {
-    maxDuration: number | null,
-    minDuration: number | null,
-    skills: SkillCount[],
-    buffs: PartyBuff[],
-    buffEffects: CombinedBuffEffect
-}
+const reverseCascade: GcdAbility = {
+    name: 'Reverse Cascade',
+    type: 'gcd',
+    potency: 340,
+    attackType: 'Weaponskill',
+    gcd: 2.50
+} as const satisfies GcdAbility;
 
-export class DncDtSim implements Simulation<DncDtSimResults, DncDtSimSettings, DncDtSimSettingsExport> {
+const fountainFall: GcdAbility = {
+    name: 'Fountainfall',
+    type: 'gcd',
+    potency: 400,
+    attackType: 'Weaponskill',
+    gcd: 2.50
+} as const satisfies GcdAbility;
+
+const saberDance: GcdAbility = {
+    name: 'Saber Dance',
+    type: 'gcd',
+    potency: 540,
+    attackType: 'Weaponskill',
+    gcd: 2.50
+} as const satisfies GcdAbility;
+
+const starfall: GcdAbility = {
+    name: 'Starfall Dance',
+    type: 'gcd',
+    potency: 600,
+    attackType: 'Weaponskill',
+    gcd: 2.50
+} as const satisfies GcdAbility;
+
+const standardFinish: GcdAbility = {
+    name: 'Standard Finish',
+    type: 'gcd',
+    potency: 720,
+    attackType: 'Weaponskill',
+    gcd: 1.50
+} as const satisfies GcdAbility;
+
+const techFinish: GcdAbility = {
+    name: 'Technical Finish',
+    type: 'gcd',
+    potency: 1200,
+    attackType: 'Weaponskill',
+    gcd: 1.50
+} as const satisfies GcdAbility;
+
+const tillana: GcdAbility = {
+    name: 'Tillana',
+    type: 'gcd',
+    potency: 440,
+    attackType: 'Weaponskill',
+    gcd: 1.50
+} as const satisfies GcdAbility;
+
+const fanDance: OgcdAbility = {
+    name: 'Fan Dance',
+    type: 'ogcd',
+    potency: 180,
+    attackType: 'Ability',
+    cooldown: {
+        time: 1
+    },
+} as const satisfies OgcdAbility;
+
+const fd3: OgcdAbility = {
+    name: 'Fan Dance III',
+    type: 'ogcd',
+    potency: 220,
+    attackType: 'Ability',
+    cooldown: {
+        time: 1
+    },
+} as const satisfies OgcdAbility;
+
+const fd4: OgcdAbility = {
+    name: 'Fan Dance IV',
+    type: 'ogcd',
+    potency: 340,
+    attackType: 'Ability',
+    cooldown: {
+        time: 1
+    },
+} as const satisfies OgcdAbility;
+
+const finishingMove: GcdAbility = {
+    name: 'Finishing Move',
+    type: 'gcd',
+    potency: 820,
+    attackType: 'Weaponskill',
+    gcd: 2.50
+} as const satisfies GcdAbility;
+
+const lastDance: GcdAbility = {
+    name: 'Last Dance',
+    type: 'gcd',
+    potency: 420,
+    attackType: 'Weaponskill',
+    gcd: 2.50
+} as const satisfies GcdAbility;
+
+const danceOfTheDawn: GcdAbility = {
+    name: 'Dance of the Dawn',
+    type: 'gcd',
+    potency: 1000,
+    attackType: 'Weaponskill',
+    gcd: 2.50
+} as const satisfies GcdAbility;
+
+const auto: AutoAttack = {
+    name: 'Auto Attack',
+    type: 'autoattack',
+    potency: 93.6,
+    attackType: 'Auto-attack'
+} as const satisfies AutoAttack;
+
+const standardStep: GcdAbility = {
+    name: 'Standard Step',
+    type: 'gcd',
+    potency: 0,
+    attackType: 'Weaponskill',
+    gcd: 1.50
+} as const satisfies GcdAbility;
+
+const techStep: GcdAbility = {
+    name: 'Technical Step',
+    type: 'gcd',
+    potency: null,
+    attackType: 'Weaponskill',
+    gcd: 1.50
+} as const satisfies GcdAbility;
+
+const stepsAction: GcdAbility = {
+    name: 'Steps',
+    type: 'gcd',
+    potency: null,
+    attackType: 'Weaponskill',
+    gcd: 1.00
+} as const satisfies GcdAbility;
+
+
+export class DncDtSim extends BaseUsageCountSim<DncDtSimResults, DncDtSimSettings> {
     readonly spec = dncDtSheetSpec;
     displayName = 'DNC Sim';
     readonly shortName = dncDtSheetSpec.stub;
     readonly manualRun = false;
 
-    readonly settings: DncDtSimSettings;
-    private readonly resultSettings: ResultSettings;
-    private readonly buffManager: BuffSettingsManager;
-    private readonly job: JobName = 'DNC';
-
-    constructor(settings?: DncDtSimSettingsExport) {
-        this.settings = this.makeDefaultSettings();
-        if (settings !== undefined) {
-            Object.assign(this.settings, settings.customSettings ?? settings);
-            this.buffManager = BuffSettingsManager.fromSaved(settings.buffConfig);
-            this.resultSettings = settings.resultSettings ?? defaultResultSettings();
-        }
-        else {
-            this.buffManager = BuffSettingsManager.defaultForJob(this.job);
-            this.resultSettings = defaultResultSettings();
-        }
-    }
-
-    async simulate(set: CharacterGearSet): Promise<DncDtSimResults> {
-        // Get enabled buffs
-        const enabledBuffs = this.buffManager.enabledBuffs;
-        // Map from duration to list of party buffs with that duration
-        const durationMap: Map<number, PartyBuff[]> = new Map();
-        enabledBuffs.forEach(buff => {
-            const duration = buff.duration;
-            const newValue = durationMap.get(duration) ?? [];
-            newValue.push(buff);
-            durationMap.set(duration, newValue);
-        });
-        // Total skills used over one cycle
-        const totals = [...this.skillsInBuffDuration(set, null)];
-        // The end state of this map is to be a map from a duration to the skills that can be used in that duration
-        // but **not** the next shorter duration.
-        const skillsDurationMap: Map<number, SkillCount[]> = new Map();
-        const durationKeys: number[] = Array.from(durationMap.keys()).filter(dur => dur);
-        const outOfBuffs = [...totals];
-        // Start with longest duration
-        let previous = outOfBuffs;
-        durationKeys.sort((a, b) => b - a);
-        for (const duration of durationKeys) {
-            // The math works by starting with the previous bucket (thus larger), and subtracting the
-            // next (thus smaller) bucket from it.
-            const thisDurationCumulative = this.skillsInBuffDuration(set, duration);
-            console.log(`Duration before ${duration} - raw ${JSON.stringify(thisDurationCumulative)}`);
-            for (const skillCount of thisDurationCumulative) {
-                const matchingSkill = previous.find(val => abilityEquals(val[0], skillCount[0]));
-                if (matchingSkill) {
-                    console.log(`Skill ${skillCount[0].name} (${duration}) = ${matchingSkill[1]} - ${skillCount[1]}`);
-                    matchingSkill[1] -= skillCount[1];
-                }
-            }
-            skillsDurationMap.set(duration, thisDurationCumulative);
-            console.log(`Duration after ${duration} - raw ${JSON.stringify(thisDurationCumulative)}`);
-            previous = thisDurationCumulative;
-        }
-        const resultBuckets: BuffWindowUsages[] = [];
-        resultBuckets.push({
-            maxDuration: null,
-            minDuration: durationKeys.length > 0 ? durationKeys[0] : null,
-            skills: [...outOfBuffs],
-            buffs: [],
-            buffEffects: noBuffEffects()
-        });
-        for (let i = 0; i < durationKeys.length; i++) {
-            const duration = durationKeys[i];
-            const minDuration = (i + 1) in durationKeys ? durationKeys[i + 1] : null;
-            const buffs = durationMap.get(duration) ?? [];
-            resultBuckets.push({
-                maxDuration: duration,
-                minDuration: minDuration,
-                skills: skillsDurationMap.get(duration),
-                buffs: buffs,
-                buffEffects: combineBuffEffects(buffs)
-            });
-        }
-        let totalPotency = 0;
-        const totalDamage: ValueWithDev = addValues(
-            ...resultBuckets.map(bucket => {
-                const buffEffects = bucket.buffEffects;
-                return addValues(...bucket.skills.flatMap(sc => {
-                    const skill = sc[0];
-                    const count = sc[1];
-                    totalPotency += skill.potency * count;
-                    // TODO: how will this handle DoTs?
-                    const dmg = abilityToDamageNew(set.computedStats, skill, buffEffects);
-                    if (dmg.directDamage) {
-                        return [multiplyFixed(dmg.directDamage, count)];
-                    }
-                    return [];
-                }));
-            })
-        );
-        const cycleTime = this.totalCycleTime(set);
-        const dps = multiplyFixed(totalDamage, 1.0 / cycleTime);
-        const pps = totalPotency / cycleTime;
-        return {
-            cycleTime: cycleTime,
-            mainDpsFull: dps,
-            mainDpsResult: applyStdDev(dps, this.resultSettings.stdDevs ?? 0),
-            totalDamage: totalDamage,
-            unbuffedPps: pps,
-            buffBuckets: resultBuckets
-        }
+    constructor(settings?: ExternalCountSettings<DncDtSimSettings>) {
+        super('DNC', settings);
     }
 
     totalCycleTime(set: CharacterGearSet): number {
-        // TODO
-        return 120;
+        const gcdTime = set.computedStats.gcdPhys(2.5);
+
+        const gcdDelay = 0.008;
+        const startFinishDelay = 0.008;
+        const stepDelay = 0.008;
+
+        const gcdCount = 45;
+        const gcdTimeWithDelay = gcdTime + gcdDelay;
+        const gcdTimeTotal = gcdCount * gcdTimeWithDelay;
+        const fixedGcdTimeTotal = 8.5;
+
+        const startFinishCount = 3;
+        const startFinishTime = startFinishCount * startFinishDelay;
+
+        const stepCount = 4;
+        const stepTime = stepCount * stepDelay;
+
+        const totalTime = gcdTimeTotal + startFinishTime + stepTime + fixedGcdTimeTotal;
+        const timeCap = 120;
+        return Math.max(totalTime, timeCap);
     }
+
 
     /**
      * Returns the number of skills that fit in a buff duration, or total if the buff duration is null.
@@ -209,17 +230,11 @@ export class DncDtSim implements Simulation<DncDtSimResults, DncDtSimSettings, D
         const gcd = set.computedStats.gcdPhys(2.5);
         // Totals, regardless of buffs
         if (buffDuration === null) {
-            return [
-                [cascade, 3.6 / gcd],
-                [fountain, 3.6 / gcd]
-            ]
+            return [[cascade, 3.6 / gcd], [fountain, 3.6 / gcd]]
         }
         // In buffs of at least 20 seconds
         else if (buffDuration >= 20) {
-            return [
-                [cascade, 0.5 / gcd],
-                [fountain, 0.5 / gcd]
-            ]
+            return [[cascade, 0.5 / gcd], [fountain, 0.5 / gcd]]
         }
         // etc
         else {
@@ -227,133 +242,7 @@ export class DncDtSim implements Simulation<DncDtSimResults, DncDtSimSettings, D
         }
     }
 
-    exportSettings(): DncDtSimSettingsExport {
-        return {
-            buffConfig: this.buffManager.exportSetting(),
-            customSettings: this.settings,
-            resultSettings: this.resultSettings
-
-        }
-    }
-
-    makeCustomConfigInterface(settings: DncDtSimSettings, updateCallback: () => void): HTMLElement | null {
-        return null;
-    }
-
-    makeConfigInterface(settings: DncDtSimSettings, updateCallback: () => void): HTMLElement {
-        const div = document.createElement("div");
-        const custom = this.makeCustomConfigInterface(settings, updateCallback);
-        if (custom) {
-            custom.classList.add('custom-sim-settings-area');
-            const section = new NamedSection('Sim-Specific Settings');
-            section.contentArea.append(custom);
-            div.appendChild(section);
-        }
-        div.appendChild(new BuffSettingsArea(this.buffManager, updateCallback));
-        div.appendChild(new ResultSettingsArea(writeProxy(this.resultSettings, updateCallback)));
-        return div;
-    }
-
-    makeToolTip?(result: DncDtSimResults): string {
-        return 'TODO';
-    }
-
-    makeResultDisplay(result: DncDtSimResults): HTMLElement {
-        // noinspection JSNonASCIINames
-        const mainResultsTable = simpleAutoResultTable({
-            "Expected DPS": result.mainDpsFull.expected,
-            "Std Deviation": result.mainDpsFull.stdDev,
-            "Expected +1σ": applyStdDev(result.mainDpsFull, 1),
-            "Expected +2σ": applyStdDev(result.mainDpsFull, 2),
-            "Expected +3σ": applyStdDev(result.mainDpsFull, 3),
-            "Unbuffed PPS": result.unbuffedPps
-        });
-        mainResultsTable.classList.add('main-results-table');
-
-        const transposedData: {
-            ability: Ability,
-            usages: Map<number, number>,
-            outOfBuffs: number,
-            total: number
-        }[] = [];
-
-        const buffDurations = new Set<number>();
-
-        result.buffBuckets.forEach(bucket => {
-            const duration: number | null = bucket.maxDuration;
-            if (duration !== null) {
-                buffDurations.add(duration);
-            }
-            bucket.skills.forEach(sc => {
-                const ability: Ability = sc[0];
-                const count: number = sc[1];
-                let abilityData = transposedData.find(datum => abilityEquals(datum.ability, ability));
-                if (!abilityData) {
-                    abilityData = {
-                        ability: ability,
-                        usages: new Map<number, number>(),
-                        outOfBuffs: 0,
-                        total: 0
-                    };
-                    transposedData.push(abilityData);
-                }
-                abilityData.total += count;
-                if (duration !== null) {
-                    abilityData.usages.set(duration, (abilityData.usages.get(duration) ?? 0) + count);
-                }
-                else {
-                    abilityData.outOfBuffs += count;
-                }
-            });
-        });
-
-        const buffDurationsSorted = Array.from(buffDurations);
-        buffDurationsSorted.sort((a, b) => (b - a));
-
-
-        const bucketsTable = new CustomTable<typeof transposedData[number]>();
-        const columns: typeof bucketsTable.columns = [{
-            shortName: "skill",
-            displayName: "Skill",
-            getter: item => item.ability,
-            renderer: (value: Ability) => {
-                return document.createTextNode(`${value.name}`);
-            }
-        }];
-        buffDurations.forEach(dur => {
-            columns.push({
-                shortName: `buff-dur-${dur}`,
-                displayName: `In ${dur}s Buffs`,
-                getter: bucket => {
-                    return bucket.usages.get(dur) ?? 0
-                },
-                renderer: value => document.createTextNode(value.toFixed(3)),
-            })
-        });
-        columns.push({
-            shortName: `out-of-buffs`,
-            displayName: `Out of Buffs`,
-            getter: bucket => {
-                return bucket.outOfBuffs
-            },
-            renderer: value => document.createTextNode(value.toFixed(3)),
-        });
-        columns.push({
-            shortName: `total`,
-            displayName: `Total`,
-            getter: bucket => {
-                return bucket.total
-            },
-            renderer: value => document.createTextNode(value.toFixed(3)),
-        });
-        bucketsTable.columns = columns;
-
-        bucketsTable.data = [new HeaderRow(), ...transposedData];
-
-        return quickElement('div', ['cycle-sim-results-table'], [mainResultsTable, bucketsTable]);
-    }
-
-    private makeDefaultSettings(): DncDtSimSettings {
+    makeDefaultSettings(): DncDtSimSettings {
         return {}
     }
 }
